backend/requirements.txt
sentence-transformers
faiss-cpu
pypdf2
numpy
openai